frontend/requirements.txt
streamlit>=1.28.0
requests>=2.25.0
pillow>=10.0.0
typing-extensions>=4.0.0 